Create counterfeitDollarFinder
Takes a series of strings representing coins on a scale and the resulting balance to determine which is counterfeit, and whether it is lighter or heavier than the real coins. There will always be 12 coins, represented by the letters A-L, and there will always be 4 coins on each side of the scale. If the strings don't provide enough information to determine the fake coin or its relative weight, the function returns '???'.  A typical string looks like the following: "ABCD EFGH even" / "ABCD EFGH up" / "ABCD EFGH down"