Create stringToPrimeNumberConverter
Takes a string of exclamation marks and question marks representing digits of a number (such that '!!' = 2, '!!???' = 23, '!????!!!!!' = 145, etc.). if the resulting number is prime, it returns that number, otherwise, it divides that number by its lowest divisor, excluding 1. It repeats this division until the result is prime, then returns it. It works through three different functions, one which converts the string to a number (getNumber(str)), one that checks if that number is prime (checkPrime(n)), and one that handles the division until prime, recursively (work(n)). All of this is housed inside a function called convert(s).