Create numberWeightSorter
takes a list of numbers (in this context, weights) and sorts them by a new property called weightWeight. The property is arrived at by summing the digits of a given number. In a case where two numbers have the same weightWeight, they are sorted not as numbers, but 'alphabetically,' as strings.